+++ docs/examples/opencensus-exporter-tracer/collector.py
@@ -19,7 +19,7 @@ from opentelemetry.exporter.opencensus.trace_exporter import (
     OpenCensusSpanExporter,
 )
 from opentelemetry.sdk.trace import TracerProvider
-from opentelemetry.sdk.trace.export import BatchExportSpanProcessor
+from opentelemetry.sdk.trace.export import BatchSpanProcessor
 
 exporter = OpenCensusSpanExporter(
     service_name="basic-service", endpoint="localhost:55678"
@@ -27,7 +27,7 @@ exporter = OpenCensusSpanExporter(
 
 trace.set_tracer_provider(TracerProvider())
 tracer = trace.get_tracer(__name__)
-span_processor = BatchExportSpanProcessor(exporter)
+span_processor = BatchSpanProcessor(exporter)
 
 trace.get_tracer_provider().add_span_processor(span_processor)
 with tracer.start_as_current_span("foo"):